add QueueService interface
